meeting page now fetches base on route
