+++ www/package.json
@@ -18,5 +18,8 @@
     "redux": "^4.1.1",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0"
+  },
+  "devDependencies": {
+    "classnames": "^2.3.1"
   }
 }
