+++ www/styles/LoadingIcon.module.css
@@ -11,7 +11,7 @@
   background-color: inherit;
   width: 100%;
   height: 100%;
-  border: solid 4px white;
+  border: solid 4px var(--textBlue);
   position: relative;
 }
 
@@ -23,7 +23,7 @@
   margin: auto;
   margin-top: 36%;
   margin-left: 47%;
-  background-color: white;
+  background-color: var(--textBlue);
   animation: spin 12s linear infinite;
 }
 
@@ -34,7 +34,7 @@
   margin: auto;
   margin-top: 31%;
   margin-left: 48.5%;
-  background-color: white;
+  background-color: var(--textBlue);
   animation: spin 1s linear infinite;
 }
 
@@ -42,9 +42,9 @@
   position:absolute;
   height: 5%;
   width: 1%;
-  background-color: white;
-  margin-top: 48%; 
-  margin-left: 49%; 
+  background-color: var(--textBlue);
+  margin-top: 48%;
+  margin-left: 49%;
   transform: translate(0, -850%);
 }
